Mess of code from debugging value iteration
For reference. Most of this will be deleted before committing to main, but there are some debugging utils that might be useful later.

Also, I implemented the state unrolling for the different format used by mdptoolbox. Might be worth getting a read on runtimes for the different implementations at some point.

For now, no plans to use mdptoolbox. Doesn't seem to add much over the few lines of code I have for value iteration.